healthcare_chatbot/requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
faiss-cpu
langchain==0.0.225
streamlit
pypdf
python-dotenv
-e .
